refactor: Exchange 추상화 레이어 추출 (#244)
- src/cryptobot/exchange/ 신설 (base.py, upbit.py)
- 기존 bot/trader.py의 pyupbit 의존을 UpbitExchange 어댑터로 이전
- bot/trader.py는 UpbitExchange를 Trader로 re-export하는 호환성 shim
- Exchange 인터페이스에 tick_size / is_market_open 추가 (주식 어댑터 대비)
- README Phase 6 추가

Related: #244
